// app-dev/src/app/scheduled-performed-surgeries/enums/status.enum.ts
export enum StatusEnum {
  stepOne = 'PENDING',
  stepTwo = 'WAITING_APPOINTMENT',
  stepThree = 'APPOINTMENT_SCHEDULED',
  stepFour = 'APPOINTMENT_HELD',
  stepFive = 'WAITING_ONSITE_APPOINTMENT',
  stepSix = 'ONSITE_APPOINTMENT_SCHEDULED',
  stepSeven = 'ONSITE_APPOINTMENT_HELD',
  stepEight = 'WAITING_SURGERY',
  stepNine = 'SURGERY_SCHEDULED',
  stepTen = 'SURGERY_PERFORMED',
  stepEleven = 'POST_SURGERY_APPOINTMENT',
  stepTwelve = 'FINISHED',
  stepThirteen = 'WAITING_PAYMENT',
}